requirements.txt
streamlit
pandas
google-generativeai
requests
openpyxl
matplotlib
scikit-learn
seaborn
